Adapt logging format to new DDNet (#53)
* Adapt logging format to new DDNet

DDNet have changed their logging format. When infclass_stats attempt to
parse it, the script crashes, immediately restarts and connects to the server
again, causing DoS.

* Add a 3-second delay for restarts

When the script fails for some reason, the loop starts it again. It
might cause DoS on the infclass server.

Do not re-launch the server immediately, instead, wait for 3 seconds.